+++ packages/babel-plugin-minify-builtins/src/index.js
@@ -1,31 +1,25 @@
 "use strict";
 
-const evaluate = require("babel-helper-evaluate-path");
 // Assuming all the static methods from below array are side effect free evaluation
 // except Math.random
 const VALID_CALLEES = ["String", "Number", "Math"];
 const INVALID_METHODS = ["random"];
 
+const newIssueUrl = "https://github.com/babel/minify/issues/new";
+
 module.exports = function({ types: t }) {
   class BuiltInReplacer {
-    constructor(program, { tdz }) {
-      this.program = program;
-      this.tdz = tdz;
+    constructor() {
       // map<expr_name, path[]>;
       this.pathsToUpdate = new Map();
     }
 
-    run() {
-      this.collect();
-      this.replace();
-    }
-
-    collect() {
+    getCollectVisitor() {
       const context = this;
 
       const collectVisitor = {
         AssignmentExpression(path) {
-          const left = path.get("left");
+          const { left } = path.node;
 
           // Should bail and not run the plugin
           // when builtin is polyfilled
@@ -42,12 +36,14 @@ module.exports = function({ types: t }) {
             return;
           }
 
+          const { node } = path;
+
           if (
-            !isComputed(path) &&
-            isBuiltin(path) &&
+            !isComputed(node) &&
+            isBuiltin(node) &&
             !getFunctionParent(path).isProgram()
           ) {
-            const expName = memberToString(path.node);
+            const expName = memberToString(node);
             addToMap(context.pathsToUpdate, expName, path);
           }
         },
@@ -59,25 +55,23 @@ module.exports = function({ types: t }) {
               return;
             }
 
+            const { node } = callee;
+
             // computed property should not be optimized
             // Math[max]() -> Math.max()
-            if (!isComputed(callee) && isBuiltin(callee)) {
-              const result = evaluate(path, { tdz: context.tdz });
-              // deopt when we have side effecty evaluate-able arguments
-              // Math.max(foo(), 1) --> untouched
-              // Math.floor(1) --> 1
-              if (result.confident && hasPureArgs(path)) {
-                path.replaceWith(t.valueToNode(result.value));
-              } else if (!getFunctionParent(callee).isProgram()) {
-                const expName = memberToString(callee.node);
-                addToMap(context.pathsToUpdate, expName, callee);
-              }
+            if (
+              !isComputed(node) &&
+              isBuiltin(node) &&
+              !getFunctionParent(callee).isProgram()
+            ) {
+              const expName = memberToString(node);
+              addToMap(context.pathsToUpdate, expName, callee);
             }
           }
         }
       };
 
-      this.program.traverse(collectVisitor);
+      return collectVisitor;
     }
 
     replace() {
@@ -100,25 +94,64 @@ module.exports = function({ types: t }) {
           for (const path of subpaths) {
             path.replaceWith(t.clone(uniqueIdentifier));
           }
+
           // hoist the created var to the top of the function scope
-          parent.get("body").unshiftContainer("body", newNode);
+          const target = parent.get("body");
+
+          /**
+           * Here, we validate a case where there is a local binding of
+           * one of Math, String or Number. Here we have to get the
+           * global Math instead of using the local one - so we do the
+           * following transformation
+           *
+           * var _Mathmax = Math.max;
+           *
+           * to
+           *
+           * var _Mathmax = (0, eval)("this").Math.max;
+           */
+          for (const builtin of VALID_CALLEES) {
+            if (target.scope.getBinding(builtin)) {
+              const prev = newNode.declarations[0].init;
+
+              if (!t.isMemberExpression(prev)) {
+                throw new Error(
+                  `minify-builtins expected a MemberExpression. ` +
+                    `Found ${prev.type}. ` +
+                    `Please report this at ${newIssueUrl}`
+                );
+              }
+
+              if (!t.isMemberExpression(prev.object)) {
+                newNode.declarations[0].init = t.memberExpression(
+                  t.memberExpression(getGlobalThis(), prev.object),
+                  prev.property
+                );
+              }
+            }
+          }
+
+          target.unshiftContainer("body", newNode);
         }
       }
     }
   }
 
+  const builtInReplacer = new BuiltInReplacer();
+
   return {
     name: "minify-builtins",
-    visitor: {
-      Program(path, { opts: { tdz = false } = {} }) {
-        const builtInReplacer = new BuiltInReplacer(path, { tdz });
-        builtInReplacer.run();
+    visitor: Object.assign({}, builtInReplacer.getCollectVisitor(), {
+      Program: {
+        exit() {
+          builtInReplacer.replace();
+        }
       }
-    }
+    })
   };
 
-  function memberToString(memberExpr) {
-    const { object, property } = memberExpr;
+  function memberToString(memberExprNode) {
+    const { object, property } = memberExprNode;
     let result = "";
 
     if (t.isIdentifier(object)) result += object.name;
@@ -128,9 +161,8 @@ module.exports = function({ types: t }) {
     return result;
   }
 
-  function isBuiltInComputed(memberExpr) {
-    const { node } = memberExpr;
-    const { object, computed } = node;
+  function isBuiltInComputed(memberExprNode) {
+    const { object, computed } = memberExprNode;
 
     return (
       computed &&
@@ -139,8 +171,8 @@ module.exports = function({ types: t }) {
     );
   }
 
-  function isBuiltin(memberExpr) {
-    const { object, property } = memberExpr.node;
+  function isBuiltin(memberExprNode) {
+    const { object, property } = memberExprNode;
 
     if (
       t.isIdentifier(object) &&
@@ -152,6 +184,68 @@ module.exports = function({ types: t }) {
     }
     return false;
   }
+
+  // Creates a segmented map that contains the earliest common Ancestor
+  // as the key and array of subpaths that are descendats of the LCA as value
+  function getSegmentedSubPaths(paths) {
+    let segments = new Map();
+
+    // Get earliest Path in tree where paths intersect
+    paths[0].getDeepestCommonAncestorFrom(
+      paths,
+      (lastCommon, index, ancestries) => {
+        // found the LCA
+        if (!lastCommon.isProgram()) {
+          let fnParent;
+          if (
+            lastCommon.isFunction() &&
+            t.isBlockStatement(lastCommon.node.body)
+          ) {
+            segments.set(lastCommon, paths);
+            return;
+          } else if (
+            !(fnParent = getFunctionParent(lastCommon)).isProgram() &&
+            t.isBlockStatement(fnParent.node.body)
+          ) {
+            segments.set(fnParent, paths);
+            return;
+          }
+        }
+        // Deopt and construct segments otherwise
+        for (const ancestor of ancestries) {
+          const fnPath = getChildFuncion(ancestor);
+          if (fnPath === void 0) {
+            continue;
+          }
+          const validDescendants = paths.filter(p => {
+            return p.isDescendant(fnPath);
+          });
+          segments.set(fnPath, validDescendants);
+        }
+      }
+    );
+    return segments;
+  }
+
+  function getChildFuncion(ancestors = []) {
+    for (const path of ancestors) {
+      if (path.isFunction() && t.isBlockStatement(path.node.body)) {
+        return path;
+      }
+    }
+  }
+
+  /**
+   * returns
+   *
+   * (0, eval)("this")
+   */
+  function getGlobalThis() {
+    return t.callExpression(
+      t.sequenceExpression([t.valueToNode(0), t.identifier("eval")]),
+      [t.valueToNode("this")]
+    );
+  }
 };
 
 function addToMap(map, key, value) {
@@ -161,68 +255,7 @@ function addToMap(map, key, value) {
   map.get(key).push(value);
 }
 
-// Creates a segmented map that contains the earliest common Ancestor
-// as the key and array of subpaths that are descendats of the LCA as value
-function getSegmentedSubPaths(paths) {
-  let segments = new Map();
-
-  // Get earliest Path in tree where paths intersect
-  paths[0].getDeepestCommonAncestorFrom(
-    paths,
-    (lastCommon, index, ancestries) => {
-      // found the LCA
-      if (!lastCommon.isProgram()) {
-        let fnParent;
-        if (
-          lastCommon.isFunction() &&
-          lastCommon.get("body").isBlockStatement()
-        ) {
-          segments.set(lastCommon, paths);
-          return;
-        } else if (
-          !(fnParent = getFunctionParent(lastCommon)).isProgram() &&
-          fnParent.get("body").isBlockStatement()
-        ) {
-          segments.set(fnParent, paths);
-          return;
-        }
-      }
-      // Deopt and construct segments otherwise
-      for (const ancestor of ancestries) {
-        const fnPath = getChildFuncion(ancestor);
-        if (fnPath === void 0) {
-          continue;
-        }
-        const validDescendants = paths.filter(p => {
-          return p.isDescendant(fnPath);
-        });
-        segments.set(fnPath, validDescendants);
-      }
-    }
-  );
-  return segments;
-}
-
-function getChildFuncion(ancestors = []) {
-  for (const path of ancestors) {
-    if (path.isFunction() && path.get("body").isBlockStatement()) {
-      return path;
-    }
-  }
-}
-
-function hasPureArgs(path) {
-  const args = path.get("arguments");
-  for (const arg of args) {
-    if (!arg.isPure()) {
-      return false;
-    }
-  }
-  return true;
-}
-
-function isComputed(path) {
-  const { node } = path;
+function isComputed(node) {
   return node.computed;
 }