+++ packages/babel-plugin-minify-dead-code-elimination/__tests__/fixtures/loops/expected.js
@@ -16,6 +16,8 @@ for (;;) {
   bar();
 }
 
+var j;
+
 // do_while
 do {
   foo();