+++ packages/babel-plugin-minify-infinity/__tests__/fixtures/expressions/actual.js
@@ -1,0 +1,10 @@
+let x = [Infinity, Infinity];
+
+let y = [
+  {
+    a: Infinity
+  },
+  {
+    a: Infinity
+  }
+];