+++ packages/babel-plugin-minify-simplify/src/index.js
@@ -120,6 +120,13 @@ module.exports = ({ types: t }) => {
         ]
       },
 
+      BinaryExpression(path) {
+        if (["!=", "=="].indexOf(path.node.operator) !== -1) {
+          undefinedToNull(path.get("left"));
+          undefinedToNull(path.get("right"));
+        }
+      },
+
       LogicalExpression: {
         exit: [
           path => {
@@ -968,47 +975,54 @@ module.exports = ({ types: t }) => {
   }
 
   function earlyReturnTransform(path) {
-    const { node } = path;
+    const block = path.get("body");
 
-    if (!t.isBlockStatement(node.body)) {
+    if (!block.isBlockStatement()) {
       return;
     }
 
-    for (let i = node.body.body.length; i >= 0; i--) {
-      const statement = node.body.body[i];
+    const body = block.get("body");
+
+    for (let i = body.length - 1; i >= 0; i--) {
+      const statement = body[i];
       if (
-        t.isIfStatement(statement) &&
-        !statement.alternate &&
-        t.isReturnStatement(statement.consequent) &&
-        !statement.consequent.argument
+        t.isIfStatement(statement.node) &&
+        !statement.node.alternate &&
+        t.isReturnStatement(statement.node.consequent) &&
+        !statement.node.consequent.argument
       ) {
-        genericEarlyExitTransform(path.get("body").get("body")[i]);
+        genericEarlyExitTransform(statement);
       }
     }
   }
 
   function earlyContinueTransform(path) {
-    const { node } = path;
+    const block = path.get("body");
 
-    if (!t.isBlockStatement(node.body)) {
+    if (!block.isBlockStatement()) {
       return;
     }
 
-    for (let i = node.body.body.length; i >= 0; i--) {
-      const statement = node.body.body[i];
+    let body = block.get("body");
+
+    for (let i = body.length - 1; i >= 0; i--) {
+      const statement = body[i];
       if (
-        t.isIfStatement(statement) &&
-        !statement.alternate &&
-        t.isContinueStatement(statement.consequent) &&
-        !statement.consequent.label
+        t.isIfStatement(statement.node) &&
+        !statement.node.alternate &&
+        t.isContinueStatement(statement.node.consequent) &&
+        !statement.node.consequent.label
       ) {
-        genericEarlyExitTransform(path.get("body").get("body")[i]);
+        genericEarlyExitTransform(statement);
       }
     }
 
+    // because we might have folded or removed statements
+    body = block.get("body");
+
     // We may have reduced the body to a single statement.
-    if (node.body.body.length === 1 && !needsBlock(node.body, node)) {
-      path.get("body").replaceWith(node.body.body[0]);
+    if (body.length === 1 && !needsBlock(block.node, path.node)) {
+      block.replaceWith(body[0].node);
     }
   }
 
@@ -1169,5 +1183,22 @@ module.exports = ({ types: t }) => {
   // is path1 an ancestor of path2
   function isAncestor(path1, path2) {
     return !!path2.findParent(parent => parent === path1);
+  }
+
+  function isPureVoid(path) {
+    return path.isUnaryExpression({ operator: "void" }) && path.isPure();
+  }
+
+  function isGlobalUndefined(path) {
+    return (
+      path.isIdentifier({ name: "undefined" }) &&
+      !path.scope.getBinding("undefined")
+    );
+  }
+
+  function undefinedToNull(path) {
+    if (isGlobalUndefined(path) || isPureVoid(path)) {
+      path.replaceWith(t.nullLiteral());
+    }
   }
 };