+++ utils/test-transform/src/test-transform.js
@@ -3,7 +3,16 @@ const babel = require("@babel/core");
 const unpad = require("unpad");
 
 function _transform(source, options) {
-  return babel.transform(unpad(source), options).code.trim();
+  // reset defaults to older babel
+  // babel-7 current beta defaults to sourceType module
+  options.sourceType = defaults(options, "sourceType", "script");
+
+  // don't use config file - babel.config.js to apply transformations
+  // this is almost never required as babel.config.js represents the
+  // project's configuration and not the config for test environment
+  options.configFile = false;
+
+  return babel.transformSync(unpad(source), options).code.trim();
 }
 
 function makeTester(
@@ -15,19 +24,23 @@ function makeTester(
   if (!Array.isArray(plugins)) {
     plugins = [plugins];
   }
+
   const thePlugin = (name, source, expected = source, babelOpts) => {
     if (typeof expected === "object") {
       babelOpts = expected;
       expected = source;
     }
+
     const { stack } = new Error();
     const options = Object.assign(
       { plugins, sourceType: "script" },
       opts,
       babelOpts
     );
+
     test(name, () => {
       const transformed = transform(source, options);
+
       try {
         check({
           transformed,
@@ -42,7 +55,9 @@ function makeTester(
       }
     });
   };
+
   thePlugin.skip = name => test.skip(name);
+
   if (excludeKeys.indexOf("inEachLine") === -1) {
     thePlugin.inEachLine = makeTester(
       plugins,
@@ -60,6 +75,7 @@ function makeTester(
       excludeKeys.concat("inEachLine")
     );
   }
+
   if (excludeKeys.indexOf("only") === -1) {
     thePlugin.only = makeTester(
       plugins,
@@ -72,6 +88,7 @@ function makeTester(
       excludeKeys.concat("only")
     );
   }
+
   return thePlugin;
 }
 
@@ -89,3 +106,11 @@ exports.snapshot = (plugins, opts) =>
       expect({ _source: source, expected: transformed }).toMatchSnapshot();
     }
   });
+
+function defaults(o, key, def) {
+  return hop(o, key) ? o[key] : def;
+}
+
+function hop(o, key) {
+  return Object.prototype.hasOwnProperty.call(o, key);
+}